+++ t1_cs/app/routers/lesson.py
@@ -4,6 +4,7 @@ from app.database import SessionLocal
 from app.services.lesson import LessonService
 from app.schemas.lesson import LessonCreate, LessonResponse, LessonUpdate
 from app.dependencies.permissions import require_professor
+from app.services.user import UserService
 
 def get_db():
     db = SessionLocal()
@@ -28,17 +29,32 @@ class LessonRouter:
             service = LessonService(db)
             return service.get_lesson_by_id(lesson_id)
 
-        @self.router.post("/", response_model=LessonResponse)
-        def create_lesson(lesson: LessonCreate, db: Session = Depends(get_db)):
+        @self.router.post("/{user_id}", response_model=LessonResponse)
+        def create_lesson(user_id: int, lesson: LessonCreate, db: Session = Depends(get_db)):
+            user_service = UserService(db)
+            user = user_service.get_user_by_id(user_id)
+            if not user:
+                raise HTTPException(status_code=404, detail="User not found")
+            require_professor(user)
             service = LessonService(db)
             return service.create_lesson(lesson)
         
-        @self.router.put("/{lesson_id}", response_model=LessonResponse)
-        def update_lesson(lesson_id: int, lesson:LessonUpdate, db: Session = Depends(get_db)):
+        @self.router.put("/{lesson_id}/{user_id}", response_model=LessonResponse)
+        def update_lesson(user_id: int, lesson_id: int, lesson:LessonUpdate, db: Session = Depends(get_db)):
+            user_service = UserService(db)
+            user = user_service.get_user_by_id(user_id)
+            if not user:
+                raise HTTPException(status_code=404, detail="User not found")
+            require_professor(user)
             service = LessonService(db)
             return service.update_lesson(lesson_id, lesson)
 
-        @self.router.delete("/{lesson_id}")
-        def delete_lesson(lesson_id: int, db: Session = Depends(get_db)):
+        @self.router.delete("/{lesson_id}/{user_id}")
+        def delete_lesson(user_id: int, lesson_id: int, db: Session = Depends(get_db)):
+            user_service = UserService(db)
+            user = user_service.get_user_by_id(user_id)
+            if not user:
+                raise HTTPException(status_code=404, detail="User not found")
+            require_professor(user)
             service = LessonService(db)
             return service.delete_lesson(lesson_id)
