+++ t1_cs/app/routers/profile.py
@@ -3,6 +3,8 @@ from sqlalchemy.orm import Session
 from app.database import SessionLocal
 from app.services.profile import ProfileService
 from app.schemas.profile import ProfileCreate, ProfileResponse
+from app.dependencies.permissions import require_admin
+from app.services.user import UserService
 
 def get_db():
     db = SessionLocal()
@@ -28,17 +30,32 @@ class ProfileRouter:
             service = ProfileService(db)
             return service.get_profile_by_id(profile_id)
 
-        @self.router.post("/", response_model=ProfileResponse)
-        def create_profile(profile: ProfileCreate, db: Session = Depends(get_db)):
+        @self.router.post("/{user_id}", response_model=ProfileResponse)
+        def create_profile(user_id: int, profile: ProfileCreate, db: Session = Depends(get_db)):
+            user_service = UserService(db)
+            user = user_service.get_user_by_id(user_id)
+            if not user:
+                raise HTTPException(status_code=404, detail="User not found")
+            require_admin(user)
             service = ProfileService(db)
             return service.create_profile(profile)
 
-        @self.router.put("/{profile_id}", response_model=ProfileResponse)
-        def update_profile(profile_id: int, profile_update: ProfileCreate, db: Session = Depends(get_db)):
+        @self.router.put("/{profile_id}/{user_id}", response_model=ProfileResponse)
+        def update_profile(profile_id: int, user_id: int,  profile_update: ProfileCreate, db: Session = Depends(get_db)):
+            user_service = UserService(db)
+            user = user_service.get_user_by_id(user_id)
+            if not user:
+                raise HTTPException(status_code=404, detail="User not found")
+            require_admin(user)
             service = ProfileService(db)
             return service.update_profile(profile_id, profile_update)
 
-        @self.router.delete("/{profile_id}", status_code=204)
-        def delete_profile(profile_id: int, db: Session = Depends(get_db)):
+        @self.router.delete("/{profile_id}/{user_id}", status_code=204)
+        def delete_profile(profile_id: int, user_id: int, db: Session = Depends(get_db)):
+            user_service = UserService(db)
+            user = user_service.get_user_by_id(user_id)
+            if not user:
+                raise HTTPException(status_code=404, detail="User not found")
+            require_admin(user)
             service = ProfileService(db)
             return service.delete_profile(profile_id)
